Fix refresh_token logic

+++ services/user-service/app/app.py
@@ -1,5 +1,5 @@
 import json
-from datetime import datetime, timedelta
+from datetime import datetime
 
 from fastapi import Depends, FastAPI, HTTPException, Cookie, Body, Response
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -17,23 +17,26 @@ users.include_routers(app)
 
 @app.post(
     "/auth/jwt/refresh",
-    response_model=users.BearerResponse,
     summary='Обновляет access и refresh токены',
     tags=['auth']
     )
 async def refresh_tokens(
-    response: Response,
     fingerprint: str = Body(..., embed=True),
     refresh_token: str = Cookie(),
     current_user=Depends(users.fastapi_users.current_user(active=True)),
     session: AsyncSession = Depends(database.get_async_session),
     ):
     refresh_token_info = await users.refreshcrud.get_refresh_token(session, refresh_token)
+    if not refresh_token_info:
+        raise HTTPException(
+            status_code=400,
+            detail='INVALID_REFRESH_SESSION'
+        )
     await users.refreshcrud.delete_refresh_token(session, refresh_token)
 
     if datetime.now() > refresh_token_info.expires_in:
         raise HTTPException(
-            status_code=500,
+            status_code=400,
             detail='TOKEN_EXPIRED'
         )
 
@@ -41,22 +44,11 @@ async def refresh_tokens(
     fingerprints = [device.fingerprint for device in devices]
     if fingerprint not in fingerprints:
         raise HTTPException(
-            status_code=500,
+            status_code=400,
             detail='INVALID_REFRESH_SESSION'
         )
 
-    tokens: users.BearerResponse = await users.update_access_and_refresh_tokens(current_user)
-    response.set_cookie(
-        key="refresh_token",
-        value=tokens.refresh_token,
-        httponly=True,
-        secure=True,
-        samesite="strict",
-        max_age=30 * 24 * 60 * 60  # 30 дней в секундах
-    )
-
-
-    return tokens
+    return await users.auth_backend.refresh_tokens(current_user)
 
 
 @app.post(
